requirements.txt
face_recognition
opencv-python-headless
numpy
python-dotenv
tk
